pyproject.toml
[tool.poetry]
name = "crismaapp"
version = "0.1.0"
description = "An simple organizational app for crisma"
authors = ["Felifelps <felipefelipe23456@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
peewee = "^3.17.1"
waitress = "^3.0.0"
python-dotenv = "^1.0.1"
bcrypt = "^4.1.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
